@@ -15,9 +15,9 @@
     "@testing-library/user-event": "^12.1.10",
     "date-fns": "^2.25.0",
     "firebase": "^9.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
