@@ -1,41 +1,29 @@
 import './App.css';
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from './Pages/Home/Home/Home';
-import Appointment from './Pages/Appointment/Appointment/Appointment';
 import Login from './Pages/Login/Login/Login';
 import Registar from './Pages/Login/Registar/Registar';
-import AuthProvider from './contexts/AuthProvider/AuthProvider';
-import PrivateRoute from './Pages/Login/PrivateRoute/PrivateRoute';
-import Dashboard from './Pages/Dashboard/Dashboard/Dashboard';
+import Appointment from './Pages/Appointment/Appointment/Appointment';
 
 function App() {
   return (
     <div className="App">
-      <AuthProvider>
-        <Router>
-          <Switch>
-            <Route exact path="/">
-              <Home />
-            </Route>
-            <Route path="/home">
-              <Home />
-            </Route>
-            <PrivateRoute path="/appointment">
-              <Appointment />
-            </PrivateRoute>
-            <PrivateRoute path="/dashboard">
-              <Dashboard />
-            </PrivateRoute>
-            <Route path="/login">
-              <Login />
-            </Route>
-            <Route path="/registar">
-              <Registar />
-            </Route>
-          </Switch>
-        </Router>
-      </AuthProvider>
-    </div>
+      <Router>
+        <Routes>
+
+          <Route path="/" element={<Home />}></Route>
+
+          <Route path="/home" element={<Home />}></Route>
+
+          <Route path="/appointment" element={<Appointment />}></Route>
+
+          <Route path="/login" element={<Login />}></Route>
+
+          <Route path="/registar" element={<Registar />}></Route>
+
+        </Routes>
+      </Router>
+    </div >
   );
 }
 
