@@ -5,20 +5,15 @@ import Appointments from '../Appointments/Appointments';
 
 const DashboardHome = () => {
 
-    const [date, setDate] = React.useState(new Date());
+
     return (
         <div>
             <Grid container spacing={2}>
                 <Grid item xs={12} md={5}>
-                    <Calender
-                        date={date}
-                        setDate={setDate}
-                    />
+
                 </Grid>
                 <Grid item xs={12} md={7}>
-                    <Appointments
-                        date={date}
-                    />
+
                 </Grid>
             </Grid>
         </div>
