@@ -1,128 +1,127 @@
-import { CircularProgress } from '@mui/material';
-import { CardElement, useElements, useStripe } from '@stripe/react-stripe-js';
-import userEvent from '@testing-library/user-event';
-import React, { useEffect, useState } from 'react';
-import useAuth from '../../../hooks/useAuth';
+// import { CircularProgress } from '@mui/material';
+// import { CardElement, useElements, useStripe } from '@stripe/react-stripe-js';
+// import React, { useEffect, useState } from 'react';
+// import useAuth from '../../../hooks/useAuth';
 
-const CheckoutForm = ({ appointment }) => {
-    const stripe = useStripe();
-    const elements = useElements();
+// const CheckoutForm = ({ appointment }) => {
+//     const stripe = useStripe();
+//     const elements = useElements();
 
-    const { user } = useAuth();
+//     const { user } = useAuth();
 
-    const { price, patientName, _id } = appointment;
-    const [error, setError] = useState('');
-    const [success, setSuccess] = useState('');
-    const [processing, setProcessing] = useState(false);
-    const [clientSecret, setClientSecret] = useState('');
-    useEffect(() => {
-        fetch('https://shielded-sea-24899.herokuapp.com/0/create-payment-intent', {
-            method: 'POST',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify({ price })
-        })
-            .then(res => res.json())
-            .then(data => setClientSecret(data.clientSecret))
-    }, [price])
+//     const { price, patientName, _id } = appointment;
+//     const [error, setError] = useState('');
+//     const [success, setSuccess] = useState('');
+//     const [processing, setProcessing] = useState(false);
+//     const [clientSecret, setClientSecret] = useState('');
+//     useEffect(() => {
+//         fetch('https://shielded-sea-24899.herokuapp.com/create-payment-intent', {
+//             method: 'POST',
+//             headers: { 'content-type': 'application/json' },
+//             body: JSON.stringify({ price })
+//         })
+//             .then(res => res.json())
+//             .then(data => setClientSecret(data.clientSecret))
+//     }, [price])
 
-    const handleSubmit = async (e) => {
+//     const handleSubmit = async (e) => {
 
-        e.preventDefault();
+//         e.preventDefault();
 
-        if (!stripe || !elements) {
-            return;
-        }
+//         if (!stripe || !elements) {
+//             return;
+//         }
 
-        const card = elements.getElement(CardElement);
+//         const card = elements.getElement(CardElement);
 
-        if (card == null) {
-            return;
-        }
-        setProcessing(true);
-        const { error, paymentMethod } = await stripe.createPaymentMethod({
-            type: 'card',
-            card
-        });
+//         if (card == null) {
+//             return;
+//         }
+//         setProcessing(true);
+//         const { error, paymentMethod } = await stripe.createPaymentMethod({
+//             type: 'card',
+//             card
+//         });
 
-        if (error) {
-            // console.log('[error]', error)
-            setError(error.message)
-            setSuccess('');
-        }
-        else {
-            setError('')
-            console.log('[paymentMethod]', paymentMethod)
-        }
+//         if (error) {
+//             // console.log('[error]', error)
+//             setError(error.message)
+//             setSuccess('');
+//         }
+//         else {
+//             setError('')
+//             console.log('[paymentMethod]', paymentMethod)
+//         }
 
-        const { paymentIntent, error: intentError } = await stripe.confirmCardPayment(
-            clientSecret,
-            {
-                payment_method: {
-                    card: card,
-                    billing_details: {
-                        name: patientName,
-                        email: user.email
-                    },
-                },
-            },
-        );
+//         const { paymentIntent, error: intentError } = await stripe.confirmCardPayment(
+//             clientSecret,
+//             {
+//                 payment_method: {
+//                     card: card,
+//                     billing_details: {
+//                         name: patientName,
+//                         email: user.email
+//                     },
+//                 },
+//             },
+//         );
 
-        if (intentError) {
-            setError(intentError.message);
-            setSuccess('')
-        } else {
-            setError('');
-            setSuccess('Your payment processed successfully');
-            console.log(paymentIntent);
-            setProcessing(false);
+//         if (intentError) {
+//             setError(intentError.message);
+//             setSuccess('')
+//         } else {
+//             setError('');
+//             setSuccess('Your payment processed successfully');
+//             console.log(paymentIntent);
+//             setProcessing(false);
 
-            const payment = {
-                amount: paymentIntent.amount,
-                created: paymentIntent.created,
-                last4: paymentMethod.card.last4,
-                transaction: paymentIntent.client_secret.slice('_secret')[0]
-            }
+//             const payment = {
+//                 amount: paymentIntent.amount,
+//                 created: paymentIntent.created,
+//                 last4: paymentMethod.card.last4,
+//                 transaction: paymentIntent.client_secret.slice('_secret')[0]
+//             }
 
-            const url = `https://shielded-sea-24899.herokuapp.com/appointment/${_id}`;
-            fetch(url, {
-                method: 'PUT',
-                headers: { 'content-type': 'application/json' },
-                body: JSON.stringify(payment)
-            })
-                .then(res => res.json())
-                .then(data => console.log(data))
+//             const url = `https://shielded-sea-24899.herokuapp.com/appointment/${_id}`;
+//             fetch(url, {
+//                 method: 'PUT',
+//                 headers: { 'content-type': 'application/json' },
+//                 body: JSON.stringify(payment)
+//             })
+//                 .then(res => res.json())
+//                 .then(data => console.log(data))
 
-        }
+//         }
 
-    }
-    return (
-        <div>
-            <h2>stripe</h2>
-            <form onSubmit={handleSubmit}>
-                <CardElement
-                    options={{
-                        style: {
-                            base: {
-                                fontSize: '16px',
-                                color: '#424770',
-                                '::placeholder': {
-                                    color: '#aab7c4',
-                                },
-                            },
-                            invalid: {
-                                color: '#9e2146',
-                            },
-                        },
-                    }}
-                />
-                {processing ? <CircularProgress /> : <button type="submit" disabled={!stripe || success}>
-                    Pay ${price}
-                </button>}
-            </form>
-            {error && <p style={{ color: 'red' }}>{error}</p>}
-            {success && <p style={{ color: 'green' }}>{success}</p>}
-        </div>
-    );
-};
+//     }
+//     return (
+//         <div>
+//             <h2>stripe</h2>
+//             <form onSubmit={handleSubmit}>
+//                 <CardElement
+//                     options={{
+//                         style: {
+//                             base: {
+//                                 fontSize: '16px',
+//                                 color: '#424770',
+//                                 '::placeholder': {
+//                                     color: '#aab7c4',
+//                                 },
+//                             },
+//                             invalid: {
+//                                 color: '#9e2146',
+//                             },
+//                         },
+//                     }}
+//                 />
+//                 {processing ? <CircularProgress /> : <button type="submit" disabled={!stripe || success}>
+//                     Pay ${price}
+//                 </button>}
+//             </form>
+//             {error && <p style={{ color: 'red' }}>{error}</p>}
+//             {success && <p style={{ color: 'green' }}>{success}</p>}
+//         </div>
+//     );
+// };
 
-export default CheckoutForm;
+// export default CheckoutForm;
