@@ -1,30 +1,20 @@
-import { CircularProgress } from '@mui/material';
-import React from 'react';
-import { Redirect, Route } from 'react-router-dom';
-import useAuth from '../../../hooks/useAuth';
+// import { CircularProgress } from '@mui/material';
+// import React from 'react';
+// import { Navigate, useLocation } from 'react-router-dom';
+// import useAuth from '../../../hooks/useAuth';
 
-const AdminRoute = ({ children, ...rest }) => {
-    const { user, admin, isLoading } = useAuth();
-    if (isLoading) {
-        return <CircularProgress />
-    }
-    return (
-        <Route
-            {...rest}
-            render={({ location }) =>
-                user.email && admin ? (
-                    children
-                ) : (
-                    <Redirect
-                        to={{
-                            pathname: "/",
-                            state: { from: location }
-                        }}
-                    />
-                )
-            }
-        />
-    );
-};
+// const AdminRoute = ({ children, ...rest }) => {
+//     const { user, admin, isLoading } = useAuth();
+//     const location = useLocation();
+//     if (isLoading) {
+//         return <CircularProgress />
+//     };
 
-export default AdminRoute;
+//     if (user.email && admin) {
+//         return children;
+//     }
+//     return <Navigate to="/" state={{ from: location }} />;
+
+// };
+
+// export default AdminRoute;
