@@ -1,30 +1,20 @@
-import { CircularProgress } from '@mui/material';
-import React from 'react';
-import { Redirect, Route } from 'react-router-dom';
-import useAuth from '../../../hooks/useAuth';
+// import { CircularProgress } from '@mui/material';
+// import React from 'react';
+// import { useLocation, Navigate } from 'react-router-dom';
+// import useAuth from '../../../hooks/useAuth';
 
-const PrivateRoute = ({ children, ...rest }) => {
-    const { user, isLoading } = useAuth();
-    if (isLoading) {
-        return <CircularProgress />
-    }
-    return (
-        <Route
-            {...rest}
-            render={({ location }) =>
-                user.email ? (
-                    children
-                ) : (
-                    <Redirect
-                        to={{
-                            pathname: "/login",
-                            state: { from: location }
-                        }}
-                    />
-                )
-            }
-        />
-    );
-};
+// const PrivateRoute = ({ children, ...rest }) => {
+//     const { user, isLoading } = useAuth();
+//     let location = useLocation();
+//     if (isLoading) {
+//         return <CircularProgress />
+//     }
+//     if (user.email) {
+//         return children;
+//     }
+//     return <Navigate to="/login" state={{ from: location }} />;
 
-export default PrivateRoute;
+
+// };
+
+// export default PrivateRoute;
